+++ src/aks-preview/HISTORY.rst
@@ -11,6 +11,10 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+
+3.0.0b10
+++++++++
+* Support to enable azure monitor profile when the sku name is automatic.
 * Vendor new SDK and bump API version to 2024-03-02-preview.
 * Add option `WindowsAnnual` to `--os-sku` for `az aks nodepool add`.
 * Add option `--enable-force-upgrade`, `--disable-force-upgrade` and `--upgrade-override-until` to `az aks upgrade`.